fix: fix: use config.skills.paths for OpenCode skill discovery and rename plugin to fractary-core.js
