+++ .opencode/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@fractary/opencode-core",
+  "version": "1.0.2",
+  "type": "module",
+  "main": "./plugins/fractary-core.js",
+  "keywords": [
+    "opencode-plugin",
+    "fractary"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/fractary/core",
+    "directory": ".opencode"
+  },
+  "license": "Apache-2.0",
+  "dependencies": {
+    "@opencode-ai/plugin": "1.3.13"
+  }
+}
